The appearance of the enemy WAVE and the corresponding base deducting blood
This time I finished creating the enemy UI and the end life values. And created corresponding scripts for the enemies and endpoints. The scenario I tried in the script was to fix the game at a certain position after reaching the end point, and if the end point life value reaches 0, the game will be terminated directly and a game over UI popup will be created later. The option of destroying the enemy UI after reaching the end is not feasible because of an unknown bug that makes all enemy UI disappear when destroyed.